Refactor of listener deduplication logic
Instead of relying on the simple file and line where the Cmd function is
invoked, we now walk the full stack until we exit this module and reach
the caller. There, we try and deduplicate again by file and line, but
also use the program counter and allow the user to give the library
hints to properly deduplicate calls.